fix: only  to run the project
