+++ src/main/java/ages/hopeful/exception/GlobalExceptionHandler.java
@@ -1,30 +1,80 @@
 package ages.hopeful.exception;
 
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.validation.FieldError;
 import org.springframework.web.bind.MethodArgumentNotValidException;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
 
+import ages.hopeful.exception.HttpException;
+
+import java.util.HashMap;
 import java.util.Map;
 
 @RestControllerAdvice
+@Slf4j
 public class GlobalExceptionHandler {
 
-  // captura qualquer HttpException (Conflict, NotFound, etc)
   @ExceptionHandler(HttpException.class)
-  public ResponseEntity<Map<String, Object>> handleHttpException(HttpException ex) {
-    return ResponseEntity
-        .status(ex.getStatus())
-        .body(Map.of("error", ex.getMessage()));
+  public ResponseEntity<?> handleHttpException(HttpException ex) {
+    return new ResponseEntity<>(new ErrorResponse(ex.getMessage(), ex.getStatus()), ex.getStatus());
   }
 
-  // captura erros de validação do @Valid
-  @ExceptionHandler(MethodArgumentNotValidException.class)
-  public ResponseEntity<Map<String, Object>> handleValidation(MethodArgumentNotValidException ex) {
-    var errors = ex.getBindingResult().getFieldErrors().stream()
-        .map(fe -> Map.of("field", fe.getField(), "message", fe.getDefaultMessage()))
-        .toList();
+  @ExceptionHandler(Exception.class)
+  public ResponseEntity<?> handleException(Exception ex) {
+    return new ResponseEntity<>(new ErrorResponse("Erro interno do servidor",
+            HttpStatus.INTERNAL_SERVER_ERROR), HttpStatus.INTERNAL_SERVER_ERROR);
+  }
 
-    return ResponseEntity.badRequest().body(Map.of("errors", errors));
+  @ExceptionHandler(MethodArgumentNotValidException.class)
+  public ResponseEntity<Map<String, Object>> handleValidationExceptions(MethodArgumentNotValidException ex) {
+    Map<String, Object> response = new HashMap<>();
+    response.put("status", "BAD_REQUEST");
+
+    Map<String, String> errors = new HashMap<>();
+
+    for (FieldError fieldError : ex.getBindingResult().getFieldErrors()) {
+      errors.put(fieldError.getField(), fieldError.getDefaultMessage());
+    }
+
+    response.put("errors", errors);
+    return new ResponseEntity<>(response, HttpStatus.BAD_REQUEST);
+  }
+
+  @ExceptionHandler(IllegalArgumentException.class)
+  public ResponseEntity<Map<String, Object>> handleIllegalArgument(IllegalArgumentException ex) {
+    Map<String, Object> response = new HashMap<>();
+    response.put("status", "BAD_REQUEST");
+    response.put("error", ex.getMessage());
+
+    return new ResponseEntity<>(response, HttpStatus.BAD_REQUEST);
+  }
+
+  public static class ErrorResponse {
+    private String message;
+    private HttpStatus status;
+
+    public ErrorResponse(String message, HttpStatus status) {
+      this.message = message;
+      this.status = status;
+    }
+
+    public String getMessage() {
+      return message;
+    }
+
+    public void setMessage(String message) {
+      this.message = message;
+    }
+
+    public HttpStatus getStatus() {
+      return status;
+    }
+
+    public void setStatus(HttpStatus status) {
+      this.status = status;
+    }
   }
 }
